backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
httpx==0.26.0
python-multipart==0.0.6
pydantic==2.5.3

# ML
scikit-learn==1.4.0
numpy==1.26.3
scipy==1.12.0

# AI
google-generativeai==0.3.2

# NLP
python-dateutil==2.8.2

# Async scheduling
apscheduler==3.10.4

# Dev
pytest==7.4.4
